Base Converter
A program everyone tries to make for fun, it converts a number from one
base to another base. I gave it two conversion options, one to store
integers and their respective places, and the other to convert those
integers to Unicode. A demonstration of this is shown in the main
method.